Move bitmasks helpers into bitmasks module
